// apps/frontend/src/shared/models/categoryModel.ts
export enum CATEGORY_MODEL {
  MUSIC = 'Музыка',
  SPORTS = 'Спорт',
  ART = 'Искусство',
  FOOD = 'Еда и напитки',
  EDUCATION = 'Образование',
  TECHNOLOGY = 'Технологии',
  TRAVEL = 'Путешествия и приключения',
  HEALTH = 'Здоровье и фитнес',
  CULTURE = 'Культура и традиции',
  CHARITY = 'Благотворительность'
}